Identify repeated words
Call out when two successively repeated words are used (e.g. "I went to
the the store."). This happens very often and it's hard to catch since
the words are fine in themselves.

This rule is not implementing the Canonical style guide. It is
identifying a syntactical error. Because of this, I chose the numbering
scheme of '000-Some-rule'. If we agree to this format for outright
mistakes, I can make an addition to the README explaining it.

Signed-off-by: Peter Matulis <peter.matulis@canonical.com>